// frontend/src/app/models/course-management/swuCurriculumData.ts
/**
 * Generated by orval v6.23.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export interface SwuCurriculumData {
    /** สถานะการใช้งาน */
    activeFlag?: boolean;
    /** custom KEY  */
    code?: string;
    /** จำนวนหน่วยกิตรวมตลอดหลักสูตร */
    creditStudy?: number;
    /** รหัสหลักสูตร สป.อว. */
    curriculumMhesiId?: string;
    /** ชื่อหลักสูตร (ภาษาอังกฤษ) */
    curriculumNameEn?: string;
    /** ชื่อหลักสูตร (ภาษาไทย) */
    curriculumNameTh?: string;
    /** PK รหัสหลักสูตร มศว Not Autorun  */
    curriculumSwuId?: string;
    /** รหัสวุฒิการศึกษา */
    degreeCode?: string;
    /** วุฒิการศึกษา (ภาษาอังกฤษ) */
    degreeNameEn?: string;
    /** วุฒิการศึกษา (ภาษาไทย) */
    degreeNameTh?: string;
    /** วุฒิการศึกษาแบบย่อ (ภาษาอังกฤษ) */
    degreeShortEn?: string;
    /** วุฒิการศึกษาแบบย่อ (ภาษาไทย) */
    degreeShortTh?: string;
    /** รหัสหน่วยงานเจ้าของหลักสูตร ref mas_department.dep_code */
    depCode?: string;
    /** รหัสระดับการศึกษา */
    eduLevelCode?: string;
    /** ชื่อระดับการศึกษา (ภาษาอังกฤษ) */
    eduLevelNameEn?: string;
    /** ชื่อระดับการศึกษา (ภาษาไทย) */
    eduLevelNameTh?: string;
    first?: number;
    /** custom กลุ่มวิชา */
    groupDetail?: string;
    /** custom ชื่อกลุ่มหมวด (ภาษาอังกฤษ) */
    groupEn?: string;
    /** custom รหัสกลุ่มหมวด */
    groupId?: string;
    /** custom ชื่อกลุ่มหมวด (ภาษาไทย) */
    groupTh?: string;
    /** รหัสสาขาวิชา */
    majorCode?: string;
    /** ชื่อสาขาวิชา (ภาษาอังกฤษ) */
    majorNameEn?: string;
    /** ชื่อสาขาวิชา (ภาษาไทย) */
    majorNameTh?: string;
    /** custom รหัสแผนการศึกษา */
    minorTypeId?: string;
    /** custom ชื่อแผนการศึกษา (ภาษาไทย) */
    minorTypeTh?: string;
    mode?: string;
    rowNum?: number;
    size?: number;
    /** จำนวนปีที่ศึกษา */
    studyYear?: number;
    /** custom รหัสวิชา (ภาษาอังกฤษ) */
    subjectCodeEn?: string;
    /** custom รหัสวิชา (ภาษาไทย) */
    subjectCodeTh?: string;
    /** custom จํานวนหน่วยกิต */
    subjectCredit?: number;
    /** custom ชื่อรายวิชา (ภาษาอังกฤษ) */
    subjectNameEn?: string;
    /** custom ชื่อรายวิชา (ภาษาไทย) */
    subjectNameTh?: string;
    /** custom ชุดวิชา */
    subjectSet?: string;
    /** custom FK (swu_subject) */
    subjectSwuId?: number;
    /** custon ids */
    subjectSwuIds?: number[];
    /** custom ประเภทวิชา (ถาษาไทย) */
    subjectTypeTh?: string;

    ownerDepNameTh?: string;
    ownerDepNameEn?: string;
}